Return error code if command is missing
Previously if no subcommand was given the application returned 0 which
indicates success. So this is the case where you run a cli command with
no arguments or an argument which resolves to a command group.

With this change it will return error code 1. This seems more useful and
conventional.

As an aside, it seems impossible to register a handler for the top-level
command. For example it's not possible to implement "ls" with clingy.